pyproject.toml
[project]
name = "mail-manage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "exchangelib>=5.5.0",
    "python-dotenv>=1.0.1",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
